+++ onprem/core.py
@@ -17,7 +17,7 @@ from langchain_openai import ChatOpenAI
 import chromadb
 import os
 import warnings
-from typing import Any, Dict, Generator, List, Optional, Tuple, Union
+from typing import Any, Dict, Generator, List, Optional, Tuple, Union, Callable
 
 # %% ../nbs/00_core.ipynb 4
 # reference: https://github.com/langchain-ai/langchain/issues/5630#issuecomment-1574222564
@@ -226,7 +226,7 @@ class LLM:
         return db
 
     def ingest(
-        self, source_directory: str, chunk_size: int = 500, chunk_overlap: int = 50
+        self, source_directory: str, chunk_size: int = 500, chunk_overlap: int = 50, ignore_fn:Optional[Callable] = None
     ):
         """
         Ingests all documents in `source_folder` into vector database.
@@ -237,12 +237,14 @@ class LLM:
         - *source_directory*: path to folder containing document store
         - *chunk_size*: text is split to this many characters by `langchain.text_splitter.RecursiveCharacterTextSplitter`
         - *chunk_overlap*: character overlap between chunks in `langchain.text_splitter.RecursiveCharacterTextSplitter`
+        - *ignore_fn*: Optional function that accepts the file path (including file name) as input and returns True
+                       if file path should not be ingested.
 
         **Returns:** `None`
         """
         ingester = self.load_ingester()
         return ingester.ingest(
-            source_directory, chunk_size=chunk_size, chunk_overlap=chunk_overlap
+            source_directory, chunk_size=chunk_size, chunk_overlap=chunk_overlap, ignore_fn=ignore_fn
         )
 
     def check_model(self):