+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Most recent releases are shown at the top. Each release shows:
 
 ### changed
 - Added `ignore_fn` argument to `LLM.ingest` to allow more control over ignoring certain files (#58)
+- Added `Ingester.get_ingested_files` to show files ingested into vector database (#59)
 
 ### fixed:
 - N/A