+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
-## 0.0.37 (TBD)
+## 0.0.37 (2024-05-23)
 
 ### new:
 - Use OnPrem.LLM with OpenAI-compatible REST APIs (#61)
@@ -18,10 +18,9 @@ Most recent releases are shown at the top. Each release shows:
   REST APIs (#62)
 - Added `ignore_fn` argument to `LLM.ingest` to allow more control over ignoring certain files (#58)
 - Added `Ingester.get_ingested_files` to show files ingested into vector database (#59)
-- If encountering a loading error when processing a file, skip and continue instead of halting (#60)
 
 ### fixed:
-- N/A
+- If encountering a loading error when processing a file, skip and continue instead of halting (#60)
 
 
 ## 0.0.36 (2024-01-16)